Парсинг в структуру
