+++ source/lexer.cpp
@@ -1,0 +1,53 @@
+#include <istream>
+
+#include "lexer.hpp"
+
+using namespace nnmcpp::parsing;
+
+Lexem::Lexem() :
+    type(LexemType::EMPTY) {}
+
+Lexem::Lexem(LexemType t) :
+    type(t) {}
+
+Lexem::Lexem(LexemType t, const std::string& v) :
+    type(t),
+    value(v) {}
+
+LexemStream Lexer::Tokenize(std::istream& in) {
+    LexemStream stream;
+    Lexem lexem;
+
+    while (in.good() && (lexem = parseLexem(in)).type != LexemType::EMPTY)
+        stream.Put(lexem);
+
+    return stream;
+}
+
+Lexem Lexer::parseLexem(std::istream& in) {
+    char c;
+    std::string s;
+
+    while (in.good()) {
+        c = in.peek();
+        switch (c) {
+            case ':':
+            case '\n':
+                if (!s.empty())
+                    return Lexem(LexemType::TEXT, s);
+
+                in.get();
+                return Lexem((c == '\n' ? LexemType::NEWLINE : LexemType::COLON), (c == '\n' ? "\n" : ":"));
+
+            default:
+                in.get();
+                s.push_back(c);
+        }
+    }
+
+    if (!s.empty())
+        return Lexem(LexemType::TEXT, s);
+
+    return Lexem(LexemType::EMPTY, "");
+}
+
